requirements.txt
numpy>=1.12.0
# python 3.5
tensorflow-gpu==1.3.0

Pillow
matplotlib